:recycle: 예약 관련 로직 리팩토링

+++ src/main/java/kr/hhplus/concert/domain/model/Queue.java
@@ -5,12 +5,13 @@ import lombok.Builder;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.UUID;
 
 @Getter
 @Builder
 public class Queue {
-    Integer id;
+    Long id;
     Long userId;
     String token;
     LocalDateTime expiredAt;
