+++ src/main/java/kr/hhplus/concert/domain/repository/SeatRepository.java
@@ -3,5 +3,5 @@ package kr.hhplus.concert.domain.repository;
 import kr.hhplus.concert.domain.model.Seat;
 
 public interface SeatRepository {
-    Seat findSeatById(Integer concertId, Integer seatId);
+    Seat findSeatById(Long concertId, Long seatId);
 }
